Removing overkill <meta> tag. May hide pages temporarily from Google Search.
https://productforums.google.com/d/msg/webmasters/0sqRrolO_Ss/igOdQIjwKdEJ

"One reason for this is that we sometimes find a non-canonical URL first. If this URL has a noindex robots meta tag, we might decide not to index anything until we crawl and index the canonical URL. Without the noindex robots meta tag (with the rel=canonical link element) we can start by indexing that URL and show it to users in search results. As soon as we crawl the canonical URL, we can change to the canonical URL instead. It's also much safer because you don't have to worry about serving different versions of the content depending on the exact URL :-)."