Uses explicit to_s when outputting post dates
In Ruby 1.9.3, this results in a binary encoding of the date when output to
YAML, and then when it is read back in, the timezone is referenced correctly.
In the prior version, the timezone was ignored and the date was seen as UTC.

Added in the GMT date explicitly, too.