Change data[author] to a string scalar
Using the minima template, data[author] is supposed to be a string scala, but the conversion turns it into a map of displayname, author_login and author_email. author_login and author_email are also stored as separate scalar attributes. The fix turns data[author] into a string scalar field of displayname, fixing the display of the author name in the minima template.